requirements.txt
customtkinter>=5.2.2
matplotlib>=3.8.0
numpy>=1.26.0
Pillow>=10.0.0
